Rename constants.ts to make purpose of file clear
